requirements.txt
fastapi
uvicorn
pydantic
requests
Pillow
python-dotenv
openai